pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "liquid-host"
version = "0.1.0"
description = "Framework for downloading and hosting Liquid AI LFM models locally"
requires-python = ">=3.10"
dependencies = [
    "transformers>=5.0.0",
    "torch>=2.1.0",
    "accelerate>=0.25.0",
    "huggingface-hub>=0.20.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "mcp>=1.0.0",
    "llama-cpp-python>=0.3.0",
    "leap-bundle>=0.9.0",
]

[project.optional-dependencies]
flash-attn = ["flash-attn>=2.5.0"]
vllm = ["vllm>=0.10.2"]
dev = ["pytest>=8.0.0", "httpx>=0.27.0"]

[project.scripts]
liquid-host = "liquid_host.cli.main:cli"

[tool.setuptools.packages.find]
include = ["liquid_host*"]
